Fix multiprocessing in Windows (Can't pickle local object)

+++ todus/client.py
@@ -1,5 +1,5 @@
 import string
-from typing import Any, Callable, Dict, Optional
+from typing import Any, Callable, Dict, Optional, Tuple
 
 import requests
 
@@ -9,7 +9,7 @@ from .util import ResultProcess, generate_token
 
 class ToDusClient:
     def __init__(
-        self, version_name: str = "0.38.34", version_code: str = "21805"
+        self, version_name: str = "0.40.16", version_code: str = "21820"
     ) -> None:
         self.version_name = version_name
         self.version_code = version_code
@@ -29,17 +29,20 @@ class ToDusClient:
         kwargs.setdefault("timeout", self.timeout)
         return self._real_request(*args, **kwargs)
 
-    def _run_task(self, task: Callable, timeout: float) -> Any:
-        self._process = ResultProcess(target=task)
-        self._process.start()
+    def _run_task(self, task: Callable, timeout: float, **kwargs) -> Any:
+        self._process = ResultProcess(target=task, **kwargs)
         try:
+            self._process.start()
             return self._process.get_result(timeout)
+        except (AttributeError, PermissionError) as ex:
+            # AttributeError: Can't pickle local object 'ToDusClient.login.<locals>.task'
+            raise ex
         finally:
             self.abort()
 
     def abort(self) -> None:
-        if self._process is not None:
-            self._process.kill()  # type: ignore
+        if self._process is not None and self._process.is_alive():
+            self._process.kill()  # type: ignore [attr-defined]
             self._process.abort()
             self._process = None
 
@@ -63,119 +66,128 @@ class ToDusClient:
             "Content-Type": "application/x-protobuf",
         }
 
+    def task_request_code(self, phone_number: str) -> None:
+        # TODO: Fix payload/data
+        headers = self.headers_auth
+        data = (
+            b"\n\n"
+            + phone_number.encode("utf-8")
+            + b"\x12\x96\x01"
+            + generate_token(150).encode("utf-8")
+        )
+        url = "https://auth.todus.cu/v2/auth/users.reserve"
+        with self.session.post(url, data=data, headers=headers) as resp:
+            resp.raise_for_status()
+
     def request_code(self, phone_number: str) -> None:
         """Request server to send verification SMS code."""
+        kwargs = {"args": (phone_number,)}
+        self._run_task(self.task_request_code, self.timeout, **kwargs)
 
-        def task() -> None:
-            headers = self.headers_auth
-            data = (
-                b"\n\n"
-                + phone_number.encode()
-                + b"\x12\x96\x01"
-                + generate_token(150).encode()
-            )
-            url = "https://auth.todus.cu/v2/auth/users.reserve"
-            with self.session.post(url, data=data, headers=headers) as resp:
-                resp.raise_for_status()
-
-        self._run_task(task, self.timeout)
+    def task_validate_code(self, phone_number: str, code: str) -> str:
+        # TODO: Fix payload/data
+        headers = self.headers_auth
+        data = (
+            b"\n\n"
+            + phone_number.encode("utf-8")
+            + b"\x12\x96\x01"
+            + generate_token(150).encode("utf-8")
+            + b"\x1a\x06"
+            + code.encode()
+        )
+        url = "https://auth.todus.cu/v2/auth/users.register"
+        with self.session.post(url, data=data, headers=headers) as resp:
+            resp.raise_for_status()
+            if b"`" in resp.content:
+                index = resp.content.index(b"`") + 1
+                return resp.content[index : index + 96].decode()
+            else:
+                return resp.content[5:166].decode()
 
     def validate_code(self, phone_number: str, code: str) -> str:
         """Validate phone number with received SMS code.
 
         Returns the account password.
         """
+        kwargs = {"args": (phone_number, code)}
+        return self._run_task(self.task_validate_code, self.timeout, **kwargs)
 
-        def task() -> str:
-            headers = self.headers_auth
-            data = (
-                b"\n\n"
-                + phone_number.encode()
-                + b"\x12\x96\x01"
-                + generate_token(150).encode()
-                + b"\x1a\x06"
-                + code.encode()
-            )
-            url = "https://auth.todus.cu/v2/auth/users.register"
-            with self.session.post(url, data=data, headers=headers) as resp:
-                resp.raise_for_status()
-                if b"`" in resp.content:
-                    index = resp.content.index(b"`") + 1
-                    return resp.content[index : index + 96].decode()
-                else:
-                    return resp.content[5:166].decode()
-
-        return self._run_task(task, self.timeout)
+    def task_login(self, phone_number: str, password: str) -> str:
+        # TODO: Fix payload/data
+        headers = self.headers_auth
+        data = (
+            b"\n\n"
+            + phone_number.encode()
+            + b"\x12\x96\x01"
+            + generate_token(150).encode("utf-8")
+            + b"\x12\x60"
+            + password.encode()
+            + b"\x1a\x05"
+            + self.version_code.encode("utf-8")
+        )
+        url = "https://auth.todus.cu/v2/auth/token"
+        with self.session.post(url, data=data, headers=headers) as resp:
+            resp.raise_for_status()
+            token = "".join([c for c in resp.text if c in string.printable])
+            return token
 
     def login(self, phone_number: str, password: str) -> str:
         """Login with phone number and password to get an access token."""
+        kwargs = {"args": (phone_number, password)}
+        return self._run_task(self.task_login, self.timeout, **kwargs)
 
-        def task() -> str:
-            headers = self.headers_auth
-            data = (
-                b"\n\n"
-                + phone_number.encode()
-                + b"\x12\x96\x01"
-                + generate_token(150).encode()
-                + b"\x12\x60"
-                + password.encode()
-                + b"\x1a\x05"
-                + self.version_code.encode()
-            )
-            url = "https://auth.todus.cu/v2/auth/token"
-            with self.session.post(url, data=data, headers=headers) as resp:
-                resp.raise_for_status()
-                token = "".join([c for c in resp.text if c in string.printable])
-                return token
+    def task_upload_file_1(self, token: str, size: int) -> Tuple[str, str]:
+        return reserve_url(token, size)
 
-        return self._run_task(task, self.timeout)
+    def task_upload_file_2(
+        self, token: str, data: bytes, up_url: str, down_url: str, timeout: float
+    ) -> str:
+        headers = {
+            "User-Agent": self.upload_ua,
+            "Authorization": f"Bearer {token}",
+        }
+        with self.session.put(
+            url=up_url, data=data, headers=headers, timeout=timeout
+        ) as resp:
+            resp.raise_for_status()
+        return down_url
 
     def upload_file(self, token: str, data: bytes, size: int = None) -> str:
         """Upload data and return the download URL."""
         if size is None:
             size = len(data)
 
-        def task1() -> tuple:
-            return reserve_url(token, size)
-
-        up_url, down_url = self._run_task(task1, self.timeout)
+        kwargs = {"args": (token, size)}
+        up_url, down_url = self._run_task(
+            self.task_upload_file_1, self.timeout, **kwargs
+        )
 
         timeout = max(len(data) / 1024 / 1024 * 20, self.timeout)
 
-        def task2() -> tuple:
-            headers = {
-                "User-Agent": self.upload_ua,
-                "Authorization": f"Bearer {token}",
-            }
-            with self.session.put(
-                url=up_url, data=data, headers=headers, timeout=timeout
-            ) as resp:
-                resp.raise_for_status()
-            return down_url
+        kwargs = {"args": (token, data, up_url, down_url, timeout)}  # type: ignore [dict-item]
+        return self._run_task(self.task_upload_file_2, timeout, **kwargs)
 
-        return self._run_task(task2, timeout)
+    def task_download_1(self, token, url) -> str:
+        return get_real_url(token, url)
+
+    def task_download_2(self, token, url, path) -> int:
+        headers = {
+            "User-Agent": self.download_ua,
+            "Authorization": f"Bearer {token}",
+        }
+        with self.session.get(url=url, headers=headers) as resp:
+            resp.raise_for_status()
+            size = int(resp.headers.get("Content-Length", 0))
+            with open(path, "wb") as file:
+                file.write(resp.content)
+            return size
 
     def download_file(self, token: str, url: str, path: str) -> int:
         """Download file URL.
 
         Returns the file size.
         """
-
-        def task1() -> str:
-            return get_real_url(token, url)
-
-        url = self._run_task(task1, self.timeout)
-
-        def task2() -> int:
-            headers = {
-                "User-Agent": self.download_ua,
-                "Authorization": f"Bearer {token}",
-            }
-            with self.session.get(url=url, headers=headers) as resp:
-                resp.raise_for_status()
-                size = int(resp.headers.get("Content-Length", 0))
-                with open(path, "wb") as file:
-                    file.write(resp.content)
-                return size
-
-        return self._run_task(task2, self.timeout)
+        kwargs = {"args": (token, url)}
+        url = self._run_task(self.task_download_1, self.timeout, **kwargs)
+        kwargs = {"args": (token, url, path)}  # type: ignore [dict-item]
+        return self._run_task(self.task_download_2, self.timeout, **kwargs)
